Configurable graph force strength and link length (#104)
* feat: Customize link length

The length of the links can be now configured via the config object. The specified length will include not only the visible segment of the link, but also the part hidden beneath the nodes it connects, which equals the half of each node' size.

* feat: Configure graph force strength

Now it's posible to customize the graph force strength, which will allow
the developer to define the gravity between the nodes. The longer the
value, the closer they will be. If, on the contrary, a big negative
value is set, the nodes will be far from each other.

* fix: Check if config.node exists before retrieving gravity

* feat: All values in D3_CONST are now configurable

The D3_CONST object has been removed, and all of the properties it had
inside are now configurable via the config object of the `Graph`
component.

All of these d3 related values are now wrapped in a new `d3` section
inside of the config object.

* refactor: Unneeded fallback line removed

* fix: Tests failing due to undefined newConfig.d3